Add feature to General Config to display fuzzy dates instead of absolute dates. Add feature to General Config to trim the leading number "0" shown on hour of day and date of month.
Affects UI dates on the Coming Episodes, Display Show, Manage > Backlog Overview, Home and History pages.

Widen the General Config tab widths and tweak, format and align some texts.

The many 'test date' variations in the General Config -> "Date and Time" drop down are useless during the month of May as both short and long styles are the same. Change the 'test date' to January 1 of the next year so short and long month styles are distinguishable.