fix episode filtering
Checking the filtering checkboxes on episode listings did
nothing on the first checkbox change, then did the opposite
of what it was supposed (i.e. checking the checkbox would
hide the episodes and vice versa).

Commit 95d7d728e00e225dd67a6f0eede10b41e5571ba8 eliminated
'return' statement in function - better to remove the code
that 'return' was skipping over anyway. This fixes the issue.