Fixed SearchFailedDownload when selecting multiple ep's as failed in displayShow
Searched for other occurrences for the search_queue.FailedQueueItem() call.
Changed segment to [segment]

The failedDownloadQueueItem always expects a single epObj in the segment. But when checking multiple episodes as "Failed" in displayShow, an array is provided. I've changed the queueItem to always iterate over it as an array. And i've made sure that when retry a single episode, it's also send to the queueItem as an array (with a single element).
I've had to adapt the queuing functionality in webserve for this.