Improve display of progress bars.
Change fill color height of progress bars in the Downloads columns of the show list page.
Change fill color height of progress bars under the images in Layout Poster of the show list page.
Remove top grey border on the up to 20% red progress bars in Layout Poster of the show list page.