Change to separate stable and develop only items in CHANGES.md.
If a commit will alter existing "stable" code then it should be documented in the main CHANGES.md section. However, if the commit relates to a change that began in develop and is not yet in stable, then it should be documented under a "develop changelog" stanza because the change is moot in the context of the "stable" end user experience.  Example, a UI feature committed on develop is not yet merged to stable, and a fix is later applied. The stable user never saw the new feature, so to document a fix to that audience is moot. However, the dev team does still need to see the change.