Fix API response header for JSON content type and the return of JSONP data.
The API should return application/json Content-Type for JSON data instead of application/html. Tornado expects a dict in order to JSON encode and send header "Content-Type:application/json". SR already encodes data plus it supports JSONP. So, the encoded string data is wrapped into a dict, and a test is added at Tornado def write() to unwrap and skip encoding, in order to set the correct content-header and also use JSONP.

Added file HACKS.txt to serve as a reminder for anyone updating the library.