Fix issue 1193
A small change to fix issue 1193 here: https://github.com/SiCKRAGETV/sickrage-issues/issues/1193

UnboundLocalError exception can be thrown at piList.append(curQueueItem) in cases where the CantUpdateException is thrown in the earlier try block around line 81. In this case, curQueueItem is never assigned a value and the list appending fails.

This change just skips the need for a temp holding variable "curQueueItem" and appends directly to the list the return from the updateShow/refreshShow methods.

Now if updateShow does tyhrow a CantUpdateException, the piList.append in the try block is never attempted and so the UnboundLocalError is avoided.