Fix NoneType object has no attribute getOverview
- Fixes AttributeError: 'NoneType' object has no attribute 'getOverview'
exception in rare cases when removing a show for which a search is still
running.

Will handle the exception and log the show's indexerID to make debugging
easier.

This will not fix the root cause, which is that the search queues will
not be properly emptied when deleting a show. This will make sure we are
able to check if this is really the cause.