webapi CMD_ShowPause doesn't update database
webapi does not update the database for the "show.pause" command till SickRage restarts. This causes the SickRage web interface to not show the "Coming Episodes" paused or not accurately. And any addons retrieving the pause data will not be correct after a webapi "show.pause" command till after a restart.

Added
showObj.saveToDB()
to lines 2391 and 2395