Fix default argument when not provided.
@MGaetan89
@miigotu

This fixes the error issue when no URI parameters are provided for "status" or "future_status".  Basically reverting to the original code for checking URI provided arguments, which I believe is/was correct based on how the webapi.py was written.  The actual default values are set later in the code if not provided by the URI parameters.  It's the same format used throughout the webapi. 

As I understand the code, the default value "None" represents no URI provided argument, for doing the comparison to the string list of possible input.  The system integer default values for "status" and "status_future" is set later in the function if no argument was supplied.

We can't use the variables (WANTED, SKIPPED, ARCHIVED, IGNORED) here since they are integers values, and "check_params" is doing comparisons against the URI arguments, and the possible values for "status" and "future_status" are text strings, not integers.

If there is a better way to do this, or I'm missing something, someone please show me.

Thanks.