New provider: www.newpct.com
Added www.newpct.com provider, prepared to seek series with Spanish
dubbing.
I've added two new options on provider's config:

- For spanish torrents: This option allows to make show searchs from
this provider ONLY if "show's config => Info Language" is marked as
Spanish. In this way we avoid search from this supplier for shows that
we have configured in VOs or other languages.

- Append identifier: Append an identifier to every episode snatched by
this provider. Usefull in combination with "Required Words" on show's
configuration if you want to download certain shows only from this
provider. In this way, if we have other providers, we avoid to snatch
torrents from other providers to prevent that episodes are downloaded in
other languages, so it is necessary that "Required Words" on show's
configuration are the same that we define here.

I hope that the changes are of your pleasure :)

CristianBB