nameparser fix
PLEASE, DON'T MERGE UNTIL YOU CAN TEST IT (It can cause a real chaos)

This code avoid to use ALL_REGEX when we know type of show to parse,
because some shows are being parsed incorrectly (anime shows being
parsed as normal shows and vice versa). This problem will affect later
to scoring system, giving invalid results.

SR matching works as follows:

- Firstly, SR has a show object with episode searched
- Secondly, for each result obtained from each provider, SR tries to
construct an object of type show
- If SR has succes building the object, SR compare it with original show
object. If they are equal gives it as valid result.

For that reason we never should send the original show object to
NameParser method (we learned it by hard way...)

This modification allow us to tell NameParser what type of show we're
parsing from findSearchResults method (generic.py) without affect others
calls to this method (there are several calls to NameParser method...)

Please be careful with this, I'm pretty sure that it's working fine but
would be great if somebody else could test it.