Improved scene number correction
Scene numbers could be corrected through the setEpisodeSceneNumbering() ajax request.
First this you need to change the value in the Scene column. But you need to use the format [season]x[episode].

I've added a regex that assumes you intent to set season 1 when only one number is entered. For ex. 5 will result in 1x5.

..added core.min.js