Show Dir is now build after indexing.
I needed to do this because anidb is giving back different show names then tvdb. As we're using tvdb as indexer, i thought it was better to use that.

but!!! For this i've made some changes to show_queue.py. When no ShowDir is provided but there is a Root_dir (forex: d:\series\anime). Then the showname is used provided by the indexer.

For this I also added a helper function that gets showname by id from indexer. As we have one indexer now, impact is limited.

But, It is possible i've overseen something. So please test this very well!.

When i'm going to refactor the other lists, i'll also implement this for imdb lists and trakt lists.