Re-factored anidb, imdb and trakt lists to use the same "techniques"
All lists should have the optional show options in top of the template. Ofcourse it would be even better if we could use one template. But for now that's not realistic. Best I can do is make sure there as much alike as possible.

core.js:
Moved the shared js functionality to the init(). Created an "action" per list, list specific code.

Made first changes to trakt trendingShows to refactor it to the same style as anidb popular anime.

Made first changes to trakt trendingShows to refactor it to the same style as anidb popular anime.

TODO: Make sure to rebase this on the consolidation of addFrom templates.