Add support for pre-executed scripts in subtitles
This patch adds new option for pre-executing scripts when searching
subtitles. Filename of media is sent as an argument for the script.

This allows one to use 3rd party utilities to enable even more sources
for subtitles. Or if for some case, one needs to do some pre-processing
before fetching/searching for subtitles.