Trying again from develop branch.
Just a smaller change now that does not interfere with the current skin, but will display shows banner from cache folder.

Personal note is that this would not need any config on/off as it fills a space in the DisplayShow page which is not used for anything and a small banner is not too big a change visually.

Hope I have understood the way this works now :>