replaced banner.png (#2309)
* replaced banner.png

Original image was stretched and generally awful. Uploaded a hi res version of the shark mascot and text.

* A few upgrades to the artwork

Some of the artworks were stretched and the shark mascot still had the watermark on it. I have re-created the artworks with the official shark mascot image and original font for the text. So it's all nice and high res.

* Delete sickrage.png

Added to wrong dir

* Delete sickrage-shark-mascot.png

Added to wrong dir

* Delete poster.png

Added to wrong dir

* Delete banner.png

Added to wrong dir

* Added updated artwork

In the correct place this time!

* Updated sickrage.png

Removed white background