Better support for show names with apostrophes
Show names with an apostrophe in the official name can sometimes exist in indexers with the apostrophe trimmed out. This change tries searching for the show name both with and without the apostrophes.

Eg.

"RuPaul's Drag Race" may be uploaded to indexers as "RuPauls Drag Race".

Made this patch to the show name helper module and confirmed that this enabled SickRage to find the show episodes on indexers where it previously could not.