Add Check for Extraction in Progress During Auto Post Processing
Prevent hanging by creating global variable that tracks whether an
extraction process was started. If the the post-processor runs again
and the process is still marked “in progress” than an error will be
logged. This may resolve issues with archives that do not throw errors
and keeps it out of the black box of the unrar binary.