Fixed authentication issue + search query
Fixed the authentication issue. Ygg has updated their page and the 'logout' text was replaced with an icon. The original script was searching for word 'logout'. Replaced this with searching for the username in the accounts page.
Started a bit of sanitizing of the search query. For example, ygg does not accept single quote ' like this one.
TODO : finish with sanitizing of search query