Fix yggtorrent login (from user:phtagn) (#4239)
* Fixed authentication issue + search query

Fixed the authentication issue. Ygg has updated their page and the 'logout' text was replaced with an icon. The original script was searching for word 'logout'. Replaced this with searching for the username in the accounts page.
Started a bit of sanitizing of the search query. For example, ygg does not accept single quote ' like this one.
TODO : finish with sanitizing of search query

* Better handling of login detection. The previous patch had a risk of misdetection in case the username was present in the page anyway (e.g. a user called href :) ).

* Following comments, amending yggtorrents.py for performance and cleanup

* An incorrect import was automatically generated by my IDE, removing

* Fix imports for isort so build passes

Signed-off-by: miigotu <miigotu@gmail.com>