Fixed logic for delete_rar
The old code would look at the subpath where the rar would extract to. Had to change the logic to look at the parent directory. Also many torrents would not return the proper result for some odd reason. This checks if a rar file exists in the folder we want to Delete. Tested  very thoroughly and works without issues.