Fix nzb splitting (#8366)
* Use request.post()'s "files" parameter in getURL

File data should be sent via request.post()'s "files" parameter.
Accept "files" parameter to getURL, and send it on to request as a POST.
(Use "files" to be consistent with the parameter name "request" uses.)

* Fix nzbSplitter so it works again.

- strip_xmlns() was broken: getchildren() was removed some time ago.
  iterator visits all element+descendants without recursion; use that.

- extraInfo[0] isn't a show object; why are we trying to use it as one?
  replace these references with obj.show which we already have anyway.

* Send file data to getURL() via "files" parameter.

Original code here said "file", but this wasn't even implemented.
Implemented file-passing parameter as "files", to be consistent with the
parameter name "request" uses, so change it here to use that name.

* Fix comparing search results to wanted episodes.

The information coming back from the search has already used scene
information to translate scene season/episode numbering into indexer
season episode numbering - we do NOT want to account for scene numbering
again, or actually matching episodes will be seen as NOT matching for
shows that are configured to use scene numbering.

Also, fix getURL() call to pass by keyword; it always should have been,
and now needs to be to still work after 'files' parameter was added.

* Don't use a single-letter variable name

---------

Co-authored-by: Rubio <Rubio9@github>