feat: automatically force a show refresh when changing anime, sports, air-by-date, or scene
This might need switched to a forced update instead of refresh.

Also, I want to use property setters and getters for most properties and get rid of DirtySetter. I might use a decorator to handle the dirty flag.
DirtySetter throws too many warnings with linters and checkers because it is non-standard practice.

We need to move refresh, update, delete, etx from sickchill.show.Show, to avoid passing an indexerid from a show object which it then loops over the show_list to get an object and then pass the object to the show_queue.